+++ src/config/additives.php
@@ -11,8 +11,8 @@
 */
 
 return [
-    "calcium" => [
-        "Canna Mono Calcium"   => [
+    "calcium"   => [
+        "Canna Mono Calcium" => [
             // Canna Mono Calcium 12% CaO
             "elements"      => [
                 "calcium" => [
@@ -23,7 +23,7 @@ return [
             "density"       => 1.283, // g/cm³
         ],
         // Ca(C₂H₃O₂)₂ · H₂O = Calciumacetat (Calciumacetat-hydrat)
-        "CaC2H3O2H2O"            => [
+        "CaC2H3O2H2O"        => [
             // Calciumacetat
             "elements"      => [
                 "calcium" => 22.7, // %
@@ -32,7 +32,7 @@ return [
             "density"       => 1.0, // g/cm³
         ],
         // Ca₃(C₆H₅O₇)₂ · 4H₂O = Calciumcitrat (Calciumcitrat-tetrahydrat)
-        "Ca3C6H5O74H2O"          => [
+        "Ca3C6H5O74H2O"      => [
             // Calciumcitrat
             "elements"      => [
                 "calcium" => 21.1, // %
@@ -40,16 +40,34 @@ return [
             "concentration" => 1.5, // %
             "density"       => 1.0, // g/cm³
         ],
-        "CACO3"                 => [
+        "CaCO3"              => [
             // Calciumcarbonat
             "elements"      => [
                 "calcium" => 40.04, // %
             ],
             "concentration" => 10.0, // %
         ],
+        "Action_Gartenkalk"  => [
+            "elements"      => [
+                "calcium"   => [
+                    "CaO" => 50.0, // %
+                ],
+                "magnesium" => 1.441, // % (5% MgCO3)
+            ],
+            "concentration" => 10.0, // %
+        ],
+        "CaO"                => [
+            // Calciumoxid
+            "elements"      => [
+                "calcium" => [
+                    "CaO" => 100.0, // %
+                ],
+            ],
+            "concentration" => 10.0, // %
+        ],
     ],
     "magnesium" => [
-        "Canna Mono Magnesium"   => [
+        "Canna Mono Magnesium" => [
             // Canna Mono Magnesium 7% MgO
             "elements"      => [
                 "magnesium" => [
@@ -59,7 +77,7 @@ return [
             "concentration" => 100.0, // %
             "density"       => 1.23, // g/cm³
         ],
-        "MgSO4"        => [
+        "MgSO4"                => [
             // Magnesiumsulfat
             "elements"      => [
                 "magnesium" => [
@@ -69,7 +87,7 @@ return [
             ],
             "concentration" => 10.0, // %
         ],
-        "C6H6MgO7"     => [
+        "C6H6MgO7"             => [
             // Magnesiumhydrogencitrat
             "elements"      => [
                 "magnesium" => [
@@ -78,10 +96,26 @@ return [
             ],
             "concentration" => 10.0, // %
         ],
-        "C12H10Mg3O14" => [
+        "C12H10Mg3O14"         => [
             // Trimagnesiumdicitrat
             "elements"      => [
-                "magnesium" => 16.0, // %
+                "magnesium" => 16.18, // %
+            ],
+            "concentration" => 10.0, // %
+        ],
+        "MgCO3"                => [
+            // Magnesiumcarbonat
+            "elements"      => [
+                "magnesium" => 28.83, // %
+            ],
+            "concentration" => 10.0, // %
+        ],
+        "MgO"                => [
+            // Magnesiumoxid
+            "elements"      => [
+                "magnesium" => [
+                    "MgO" => 100.0, // %
+                ],
             ],
             "concentration" => 10.0, // %
         ],